Simplification, get rid of a member variable.
Change-Id: I5a4e72837302c74a685e2e3a8e221345fb1871d7
Chronus: Location provider fixes and cleanup

This commit switches Chronus to only use the Passive location
provider and adds the FINE location permission.  It also includes
some general cleanup and better handling of the setting of a
custom location, fixing a NPE and ensuring the summary text is
updated correctly.

Change-Id: Ic3e870eefb66b0b81fa45ab9da6b3a8b04e63334

committed by DvTonder